Update Chroma collection name and add optional URL and collection metadata
- Update `collectionName` to use the value from the environment variable `CHROMA_COLLECTION`
- Add an optional `url` parameter with a default value of `'http://localhost:8000'`
- Add an optional `collectionMetadata` object to specify the distance method of the embedding space

---

This commit updates the code in 02_RAG.js. It changes the `collectionName` property of the `vectorStore` object to use the value from the environment variable `CHROMA_COLLECTION`. Additionally, it adds an optional `url` parameter with a default value of `'http://localhost:8000'`, and an optional `collectionMetadata` object that can be used to specify the distance method of the embedding space.